fix: add goal abort when return
