+++ src/pdf_beamtime/src/beamtime_fidpose_entrypoint.cpp
@@ -1,0 +1,53 @@
+/*Copyright 2024 Brookhaven National Laboratory
+BSD 3 Clause License. See LICENSE.txt for details.*/
+#include <pdf_beamtime/pdf_beamtime_fidpose_server.hpp>
+
+int main(int argc, char * argv[])
+{
+  rclcpp::init(argc, argv);
+
+  // Create a ROS logger for main scope
+  auto const logger = rclcpp::get_logger("hello_moveit");
+  using namespace std::chrono_literals;
+
+  // Create a node for synchronously grabbing params
+  auto parameter_client_node = rclcpp::Node::make_shared("param_client");
+  auto parent_parameters_client =
+    std::make_shared<rclcpp::SyncParametersClient>(parameter_client_node, "move_group");
+  // Boiler plate wait block
+  while (!parent_parameters_client->wait_for_service(1s)) {
+    if (!rclcpp::ok()) {
+      RCLCPP_ERROR(
+        logger, "Interrupted while waiting for the service. Exiting.");
+      return 0;
+    }
+    RCLCPP_INFO(logger, "move_group service not available, waiting again...");
+  }
+  // Get robot config parameters from parameter server
+  auto parameters = parent_parameters_client->get_parameters(
+    {"robot_description_semantic",
+      "robot_description"});
+
+  // Set node parameters using NodeOptions
+  rclcpp::NodeOptions node_options;
+  node_options.automatically_declare_parameters_from_overrides(true);
+  node_options.parameter_overrides(
+  {
+    {"robot_description_semantic", parameters[0].value_to_string()},
+    {"robot_description", parameters[1].value_to_string()}
+  });
+
+  rclcpp::executors::MultiThreadedExecutor executor;
+
+  auto beamtime_server = std::make_shared<PdfBeamtimeFidPoseServer>(
+    "ur_arm",
+    node_options,
+    "pdf_beamtime_fidpose_action_server");
+
+  executor.add_node(beamtime_server->getNodeBaseInterface());
+  executor.add_node(beamtime_server->getInterruptNodeBaseInterface());
+  executor.spin();
+
+  rclcpp::shutdown();
+  return 0;
+}
