+++ src/pdf_beamtime/src/pdf_beamtime_fidpose_server.cpp
@@ -85,6 +85,8 @@ void PdfBeamtimeFidPoseServer::execute(
         if (pickup_pose_saved) {
           fsm_results = run_return_fsm(fidpose_goal);
         } else {
+          results->success = false;
+          goal_handle->abort(results);
           RCLCPP_ERROR(node_->get_logger(), "Pickup poses are not set!");
           return;
         }
@@ -574,53 +576,4 @@ void PdfBeamtimeFidPoseServer::execute_cleanup()
   reset_fsm();
   inner_state_machine_->set_internal_state(Internal_State::RESTING);
   RCLCPP_INFO(node_->get_logger(), "Cleanup is complete");
-}
-
-int main(int argc, char * argv[])
-{
-  rclcpp::init(argc, argv);
-
-  // Create a ROS logger for main scope
-  auto const logger = rclcpp::get_logger("hello_moveit");
-  using namespace std::chrono_literals;
-
-  // Create a node for synchronously grabbing params
-  auto parameter_client_node = rclcpp::Node::make_shared("param_client");
-  auto parent_parameters_client =
-    std::make_shared<rclcpp::SyncParametersClient>(parameter_client_node, "move_group");
-  // Boiler plate wait block
-  while (!parent_parameters_client->wait_for_service(1s)) {
-    if (!rclcpp::ok()) {
-      RCLCPP_ERROR(
-        logger, "Interrupted while waiting for the service. Exiting.");
-      return 0;
-    }
-    RCLCPP_INFO(logger, "move_group service not available, waiting again...");
-  }
-  // Get robot config parameters from parameter server
-  auto parameters = parent_parameters_client->get_parameters(
-    {"robot_description_semantic",
-      "robot_description"});
-
-  // Set node parameters using NodeOptions
-  rclcpp::NodeOptions node_options;
-  node_options.automatically_declare_parameters_from_overrides(true);
-  node_options.parameter_overrides(
-  {
-    {"robot_description_semantic", parameters[0].value_to_string()},
-    {"robot_description", parameters[1].value_to_string()}
-  });
-
-  rclcpp::executors::MultiThreadedExecutor executor;
-
-  auto beamtime_server = std::make_shared<PdfBeamtimeFidPoseServer>(
-    "ur_arm",
-    node_options);
-
-  executor.add_node(beamtime_server->getNodeBaseInterface());
-  executor.add_node(beamtime_server->getInterruptNodeBaseInterface());
-  executor.spin();
-
-  rclcpp::shutdown();
-  return 0;
 }
